last exercise

+++ controllers/blogs.js
@@ -1,21 +1,31 @@
 const BlogsRouter=require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const middleware = require('../utils/middleware')
+const logger = require('../utils/logger')
 
 BlogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user','username name id')
   response.json(blogs)
 })
 
-BlogsRouter.post('/', async (request, response) => {
+BlogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
-  const user = await User.findById(body.userId)
+  logger.info('🚀 ~ BlogsRouter.post ~ body:', body)
+  let user = request.user
+  logger.info('🚀 ~ BlogsRouter.post ~ user:', user)
+
+  if (!user) {
+    user = await User.findOne()
+  }
+  logger.info('🚀 ~ BlogsRouter.post ~ user-afterfindOne():', user)
   const blog = new Blog({
     title: body.title,
     url: body.url,
     author: body.author,
-    user: user ? user.id: null
+    user: user.id
   })
+  logger.info('BLOG TO POST: ', blog)
   const savedBlog = await blog.save()
   if (user) {
     user.blogs = user.blogs.concat(savedBlog._id)
@@ -33,9 +43,22 @@ BlogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-BlogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+BlogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+  console.log('🚀 ~ BlogsRouter.delete ~ blog:', blog)
+  console.log('🚀 ~ BlogsRouter.delete ~ user:', user._id.toString())
+  console.log('🚀 ~ BlogsRouter.delete ~ blog.user:', blog.user.toString())
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(403).json({ error: 'user not allowed to perform that action' })
+  }
 })
 
 BlogsRouter.put('/:id', async (request, response) => {
